LayoutTests/editing/inserting/insert-br-quoted-007-expected.txt

CONSOLE MESSAGE: SUCCESS! newline inserted AFTER quoted content

Line of text
